Fixed Playing Pieces from Hand
Fully implemented playing pieces from a player's hand. Fixed a couple of Logic classes to handle pieces in the hand alongside the existing functionality for pieces on the board.
Also restricted selection of pieces to the active player.

Note: I changed the Move(Piece piece, PieceLocation end) constructor as it did not function with Pieces located in a player's hand. It was previously getting the player of the piece at a specified location, but all pieces in hand are at the same location so it was instead just getting the player at the top of the list. It now gets the player based on the PlayerType of the Piece. If this causes any issues, I'm sorry 😭