Fixed Some Bugs and Added Check Visual
Added a visual change to the board when a king is in check. It will change the color of the square below it to a dark red to notify they user it is in danger.
Fixed a bug with the timer running multiple times when implemented with server functionality
Fixed a bug in the move calculator that allowed the King to move into check when capturing an opponent's piece (essentially ignored protected pieces and charged into battle like a GigaChad)